feat: update WeekView component to use stable keys and improve accessibility
feat: re-export additional UI components in index.tsx

feat: create Label component with improved styling

feat: create Select component with enhanced styling and functionality

fix: remove unnecessary import in useSchedulingData hook

refactor: update tsconfig.json to remove rootDir and clean up paths

chore: update vite.config.ts to exclude @segment/analytics-node

fix: improve performance comparison logic in package-manager-config.ts

feat: add SignUpFormSimple component for user registration

feat: create Checkbox component for form handling
