+++ apps/web/tsconfig.json
@@ -38,7 +38,6 @@
     "removeComments": false,
     "incremental": true,
     "composite": true,
-    "rootDir": "./src",
     "baseUrl": ".",
     "tsBuildInfoFile": "./.tsbuildinfo",
     "types": ["vite/client", "node"],
@@ -67,8 +66,7 @@
     "vite-env.d.ts",
     "src/**/*.ts",
     "src/**/*.tsx",
-    "src/**/*.d.ts",
-    "../../packages/types/src"
+    "src/**/*.d.ts"
   ],
   "exclude": [
     "node_modules",
